aerc_project/requirements.txt
Django==5.0.1
matplotlib==3.7.2
pycryptodome==3.20.0
pydotplus==2.0.2
pyparsing==3.0.9
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.9.0
requests==2.31.0
django-apscheduler==0.6.2
django-extensions==3.2.3